Create Remove Duplicates from Sorted Array
This problem is broken, so don't try and solve it. They ask for a single number, representing the length of an array without duplicates, the method itself returning a single integer. But the judge checks the elements that are in the given length of the array, that you need to give. It doesn't make sense that the method's return type is int and not int[]. Either way, here's the solution to the problem they wanted, but can't check.